+++ python/sierrapy/commands/__init__.py
@@ -1,0 +1,7 @@
+from .cli import cli
+from . import fasta  # noqa
+from . import mutations  # noqa
+from . import patterns  # noqa
+from . import introspection  # noqa
+
+__all__ = ['cli']