+++ python/sierrapy/cmds.py
@@ -5,5 +5,9 @@ from __future__ import print_function
 from sierrapy.commands import cli
 
 
-if __name__ == '__main__':
+def main():
     cli(obj={})
+
+
+if __name__ == '__main__':
+    main()