+++ python/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.3 (unstable)
   - expose HIValg parameters in `pattern_analysis` method
+  - refactor command line tool mainly by replacing `argparse` with `click`
 
 ## 0.1.2 (2016-12-19)